Create more modular API using templated classes
Types of arguments are no longer specified using `OptionType` enum.
Instead, there are separate classes (templates), that handle parsing of
a particular type. The parser interfaces with them using non-template
base class pointer. This allows the user to easily add other types to
the parser. However, it is now up to the user to store all the parsed
options. The `ArgParser` class stores only the unrecognized options.